+++ src/sections/Projects.js
@@ -4,7 +4,7 @@ import { makeStyles } from '@material-ui/styles'
 import Container from 'components/Container'
 import { Typography, Grid } from '@material-ui/core'
 import Headline from 'components/Headline'
-import Card from 'components/Card'
+import Project from 'components/Project'
 import projects from 'data/projects'
 
 const useStyles = makeStyles(theme => ({
@@ -33,7 +33,7 @@ const Projects = props => {
       >
         {projects.map((project, i) => (
           <Grid item key={i} xs={isMobile ? 12 : 6}>
-            <Card project={project} />
+            <Project project={project} />
           </Grid>
         ))}
       </Grid>
